added dashboard object and rendering
